final functioning code for US3

+++ lib/src/main/java/edu/ucsd/cse110/habitizer/lib/domain/Routine.java
@@ -94,8 +94,8 @@ public class Routine implements Serializable {
         );
     }
 
-    public void swapTasks(Integer id1, Integer id2){
-        dataSource.swapSortOrders(id1, id2); //added to communicate to task data source
+    public void swapTasks(long id1, long id2){
+        dataSource.swapSortOrders(Math.toIntExact(id1), Math.toIntExact(id2)); //added to communicate to task data source
     }
 
     public int getGoalTime(){
